+++ components/widget_assembler_similar_result_preview/widget_comic_info/comic_info_presenter.py
@@ -2,6 +2,7 @@ import os
 
 import lzytools.file
 from PySide6.QtCore import QObject
+from PySide6.QtWidgets import QMessageBox
 
 from common import function_file
 from common.class_comic import FileType, ComicInfo
@@ -19,11 +20,16 @@ class ComicInfoPresenter(QObject):
         self.model = model
 
         self.comic_info: ComicInfo = None  # 显示的漫画的漫画信息类
+        self.is_reconfirm_before_delete = True  # 删除前是否需要再次确认
 
         # 绑定信号
         self.viewer.OpenPath.connect(self.open_path)
         self.viewer.RefreshInfo.connect(self.refresh_info)
         self.viewer.Delete.connect(self.delete_comic)
+
+    def is_reconfirm_before_delete(self, is_reconfirm: bool):
+        """设置是否删除前再次确认"""
+        self.is_reconfirm_before_delete = is_reconfirm
 
     def set_comic_info(self, comic_info: ComicInfo):
         """设置需要显示的漫画的漫画信息类"""
@@ -39,9 +45,23 @@ class ComicInfoPresenter(QObject):
         self.set_comic_info(comic_info)
 
     def delete_comic(self):
-        """删除漫画"""
-        path = self.comic_info.filepath
-        lzytools.file.delete(path, send_to_trash=True)
+        """删除文件"""
+        is_delete = True
+        if self.is_reconfirm_before_delete:
+            reply = QMessageBox.question(
+                self.viewer,
+                '确认删除',
+                '是否删除本地漫画（到回收站）',
+                QMessageBox.Yes | QMessageBox.No,  # 提供“是”和“否”两个按钮
+                QMessageBox.No  # 默认聚焦在“否”按钮上
+            )
+
+            if reply == QMessageBox.No:
+                is_delete = False
+
+        if is_delete:
+            path = self.comic_info.filepath
+            lzytools.file.delete(path, send_to_trash=True)
 
     def _show_comic_info(self):
         """在viewer上显示漫画信息"""
